chore: 🤖 restructure tests
